move tests for commands into their own files for better visibility
