+++ tests/software_tests/command_integration.rs
@@ -6,398 +6,7 @@ use STS1_EDU_Scheduler::communication::{CSBIPacket::*, CommunicationError};
 mod common;
 use common::ComEvent::*;
 
-use self::common::ComEvent;
-
 type TestResult = Result<(), Box<dyn std::error::Error>>;
-
-///// STORE ARCHIVE
-
-#[test]
-fn store_archive() -> TestResult {
-    // Define what should happen during communication. How this should look is defined in the PDD
-    let packets = vec![
-        COBC(DATA(vec![0x01, 0x00, 0x00])), // COBC sends Store Archive Command (0x01 -> Header, [0x00, 0x00] -> Program Id)
-        EDU(ACK),                           // EDU acknowledges packet integrity
-        COBC(DATA(fs::read("./tests/student_program.zip")?)), // COBC sends the archive
-        EDU(ACK),                           // EDU acknowledges packet integrity
-        COBC(EOF),                          // COBC signals end of packets
-        EDU(ACK),                           // EDU signals successful Store Archive
-    ];
-
-    // Setup testing environment
-    let (mut com, mut exec) = common::prepare_handles(packets, "0"); // construct handles for process_command
-
-    // Run actual test
-    command::handle_command(&mut com, &mut exec); // test the command processing
-
-    // Check if all packets haven been sent/received
-    assert!(com.is_complete());
-
-    // Perform further checks
-    assert_eq!(
-        0,
-        std::process::Command::new("diff") // check wether the archive was stored correctly
-            .args(["-yq", "--strip-trailing-cr", "tests/test_data", "archives/0"])
-            .status()?
-            .code()
-            .unwrap()
-    );
-
-    // Cleanup testing environment
-    common::cleanup("0");
-    Ok(())
-}
-
-#[test]
-fn stopped_store() -> TestResult {
-    let packets = vec![
-        COBC(DATA(vec![0x01, 0x04, 0x00])), // Store Archive with ID 0
-        EDU(ACK),
-        COBC(DATA(fs::read("./tests/student_program.zip")?)),
-        EDU(ACK),
-        COBC(DATA(vec![0, 1, 2, 3])),
-        EDU(ACK),
-        COBC(STOP),
-    ];
-
-    let (mut com, mut exec) = common::prepare_handles(packets, "4");
-
-    command::handle_command(&mut com, &mut exec);
-
-    assert!(!std::path::Path::new("./archives/4").exists());
-
-    common::cleanup("4");
-    Ok(())
-}
-
-#[test]
-fn invalid_crc() -> TestResult {
-    let mut bytes = fs::read("./tests/student_program.zip")?;
-    let packets = vec![
-        COBC(DATA(vec![1, 14, 0])),
-        EDU(ACK),
-        COBC(DATA(bytes.drain(0..20).collect())),
-        EDU(ACK),
-        COBC_INVALID(vec![0x8b, 5, 0, 0, 0, 0, 0, 0, 0, 10, 10, 10]),
-        EDU(NACK),
-        COBC(DATA(bytes)),
-        EDU(ACK),
-        COBC(EOF),
-        EDU(ACK),
-    ];
-    let (mut com, mut exec) = common::prepare_handles(packets, "14");
-
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    assert_eq!(
-        0,
-        std::process::Command::new("diff") // check wether the archive was stored correctly
-            .args(["-yq", "--strip-trailing-cr", "tests/test_data", "archives/14"])
-            .status()?
-            .code()
-            .unwrap()
-    );
-
-    common::cleanup("14");
-    Ok(())
-}
-
-///// EXECUTE ROGRAM
-
-#[test]
-fn execute_program_normal() -> TestResult {
-    let packets = vec![
-        COBC(DATA(vec![0x02, 0x01, 0x00, 0x00, 0x00, 0x02, 0x00])), // Execute Program ID 1, Queue ID 0, Timeout 2s
-        EDU(ACK),
-        EDU(ACK),
-    ];
-    common::prepare_program("1");
-    let (mut com, mut exec) = common::prepare_handles(packets, "1");
-
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    std::thread::sleep(std::time::Duration::from_millis(500));
-    let mut res = String::new();
-    std::fs::File::open("./archives/1/results/0")?.read_to_string(&mut res)?;
-
-    assert_eq!(res.replace('\r', ""), *"Some test results\nWith multiple lines\n".to_string());
-
-    common::cleanup("1");
-    Ok(())
-}
-
-#[test]
-fn execute_program_infinite() -> TestResult {
-    let packets = vec![
-        COBC(DATA(vec![0x02, 0x02, 0x00, 0x01, 0x00, 0x01, 0x00])), // Execute Program ID 2, Queue ID 1, Timeout 1s
-        EDU(ACK),
-        EDU(ACK),
-    ];
-    common::prepare_program("2");
-    let (mut com, mut exec) = common::prepare_handles(packets, "2");
-
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    std::thread::sleep(std::time::Duration::from_millis(1300));
-    assert!(!exec.lock().unwrap().status_queue.is_empty()?);
-
-    common::cleanup("2");
-    Ok(())
-}
-
-#[test]
-fn execute_missing_program() -> TestResult {
-    let packets = vec![COBC(DATA(vec![2, 11, 0, 0, 0, 1, 0])), EDU(ACK), EDU(NACK)];
-    let (mut com, mut exec) = common::prepare_handles(packets, "12");
-
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    common::cleanup("12");
-    Ok(())
-}
-
-///// STOP PROGRAM
-
-#[test]
-fn stop_program() -> TestResult {
-    let packets = vec![
-        COBC(DATA(vec![0x02, 0x03, 0x00, 0x01, 0x00, 0x0a, 0x00])), // Execute Program 3, Queue 1, Timeout 10s
-        EDU(ACK),
-        EDU(ACK),
-        SLEEP(std::time::Duration::from_secs(1)),
-        COBC(DATA(vec![0x03])),
-        EDU(ACK),
-        EDU(ACK),
-    ];
-    common::prepare_program("3");
-    let (mut com, mut exec) = common::prepare_handles(packets, "3");
-
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-    assert!(!exec.lock().unwrap().status_queue.is_empty()?);
-
-    common::cleanup("3");
-    Ok(())
-}
-
-#[test]
-fn stop_no_running_program() -> TestResult {
-    let packets = vec![COBC(DATA(vec![3])), EDU(ACK), EDU(ACK)];
-    let (mut com, mut exec) = common::prepare_handles(packets, "11");
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-    Ok(())
-}
-
-///// GET STATUS
-
-#[test]
-fn get_status_none() -> TestResult {
-    let packets = vec![COBC(DATA(vec![4])), EDU(ACK), EDU(DATA(vec![0])), COBC(ACK)];
-
-    let (mut com, mut exec) = common::prepare_handles(packets, "5");
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    common::cleanup("5");
-    Ok(())
-}
-
-#[test]
-fn get_status_finished() -> TestResult {
-    let packets = vec![
-        COBC(DATA(vec![0x02, 0x06, 0x00, 0x00, 0x00, 0x01, 0x00])), // Execute Program 6, Queue 0, Timeout 1s
-        EDU(ACK),
-        EDU(ACK),
-        SLEEP(std::time::Duration::from_millis(500)),
-        COBC(DATA(vec![4])), // Get Status
-        EDU(ACK),
-        EDU(DATA(vec![1, 6, 0, 0, 0, 0])), // Program Finished
-        COBC(ACK),
-        COBC(DATA(vec![4])), // Get Status
-        EDU(ACK),
-        EDU(DATA(vec![2, 6, 0, 0, 0])), // Result Ready
-        COBC(ACK),
-    ];
-
-    common::prepare_program("6");
-    let (mut com, mut exec) = common::prepare_handles(packets, "6");
-
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    common::cleanup("6");
-    Ok(())
-}
-
-#[test]
-fn get_status_priority_for_status() -> TestResult {
-    let packets = vec![
-        COBC(DATA(vec![2, 15, 0, 0, 0, 2, 0])),
-        EDU(ACK),
-        EDU(ACK),
-        SLEEP(std::time::Duration::from_millis(500)),
-        COBC(DATA(vec![4])),
-        EDU(ACK),
-        EDU(DATA(vec![1, 15, 0, 0, 0, 0])),
-        COBC(ACK),
-        COBC(DATA(vec![2, 15, 0, 0, 0, 2, 0])),
-        EDU(ACK),
-        EDU(ACK),
-        SLEEP(std::time::Duration::from_millis(500)),
-        COBC(DATA(vec![4])),
-        EDU(ACK),
-        EDU(DATA(vec![1, 15, 0, 0, 0, 0])),
-        COBC(ACK),
-    ];
-    common::prepare_program("15");
-    let (mut com, mut exec) = common::prepare_handles(packets, "15");
-
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    common::cleanup("15");
-    Ok(())
-}
-
-///// RETURN RESULT
-
-#[test]
-fn return_result() -> TestResult {
-    let packets = vec![
-        COBC(DATA(vec![0x02, 0x07, 0x00, 0x03, 0x00, 0x01, 0x00])), // Execute Program 7, Queue 0, Timeout 1s
-        EDU(ACK),
-        EDU(ACK),
-        SLEEP(std::time::Duration::from_millis(500)),
-        COBC(DATA(vec![4])), // Get Status
-        EDU(ACK),
-        EDU(DATA(vec![1, 7, 0, 3, 0, 0])), // Program Finished
-        COBC(ACK),
-        COBC(DATA(vec![4])), // Get Status
-        EDU(ACK),
-        EDU(DATA(vec![2, 7, 0, 3, 0])), // Result Ready
-        COBC(ACK),
-        COBC(DATA(vec![5])),
-        EDU(ACK),
-        ACTION(Box::new(|bytes| {
-            std::fs::File::create("tests/tmp/7.zip").unwrap().write(&bytes).unwrap();
-        })),
-        COBC(ACK),
-        EDU(EOF),
-        COBC(ACK),
-    ];
-
-    common::prepare_program("7");
-    let (mut com, mut exec) = common::prepare_handles(packets, "7");
-
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    std::process::Command::new("unzip")
-        .current_dir("./tests/tmp")
-        .arg("-o")
-        .arg("7.zip")
-        .status()?;
-
-    assert_eq!(std::fs::read("tests/tmp/3")?, vec![0xde, 0xad]);
-    assert!(std::fs::read("tests/tmp/7_3.log").is_ok());
-
-    common::cleanup("7");
-    Ok(())
-}
-
-#[test]
-fn truncate_result() -> TestResult {
-    let packets = vec![
-        COBC(DATA(vec![2, 8, 0, 5, 0, 5, 0])), // Execute Program 8, Queue 5, Timeout 2s
-        EDU(ACK),
-        EDU(ACK),
-        SLEEP(std::time::Duration::from_millis(3000)),
-        COBC(DATA(vec![4])),
-        EDU(ACK),
-        EDU(DATA(vec![1, 8, 0, 5, 0, 0])),
-        COBC(ACK),
-    ];
-
-    common::prepare_program("8");
-    let (mut com, mut exec) = common::prepare_handles(packets, "8");
-
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    assert!(fs::File::open("./data/8_5.zip")?.metadata()?.len() < 1_001_000);
-
-    common::cleanup("8");
-    Ok(())
-}
-
-#[test]
-fn stopped_return() -> TestResult {
-    let packets = vec![
-        COBC(DATA(vec![2, 9, 0, 5, 0, 3, 0])),
-        EDU(ACK),
-        EDU(ACK),
-        SLEEP(std::time::Duration::from_millis(3000)),
-        COBC(DATA(vec![4])),
-        EDU(ACK),
-        EDU(DATA(vec![1, 9, 0, 5, 0, 0])),
-        COBC(ACK),
-        COBC(DATA(vec![4])),
-        EDU(ACK),
-        EDU(DATA(vec![2, 9, 0, 5, 0])),
-        COBC(ACK),
-        COBC(DATA(vec![5])),
-        EDU(ACK),
-        ANY,
-        COBC(ACK),
-        ANY,
-        COBC(STOP),
-        COBC(DATA(vec![4])),
-        EDU(ACK),
-        EDU(DATA(vec![2, 9, 0, 5, 0])),
-        COBC(ACK),
-    ];
-    common::prepare_program("9");
-    let (mut com, mut exec) = common::prepare_handles(packets, "9");
-
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    assert!(fs::File::open("./data/9_5.zip").is_ok());
-
-    common::cleanup("9");
-    Ok(())
-}
-
-#[test]
-fn no_result_ready() -> TestResult {
-    let packets = vec![COBC(DATA(vec![5])), EDU(ACK), EDU(NACK)];
-    let (mut com, mut exec) = common::prepare_handles(packets, "10");
-
-    command::handle_command(&mut com, &mut exec);
-    assert!(com.is_complete());
-
-    common::cleanup("10");
-    Ok(())
-}
 
 #[test]
 fn invalid_packets_from_cobc() -> TestResult {
