+++ tests/tests.rs
@@ -4,5 +4,17 @@ mod command_integration;
 #[path = "software_tests/communication_tests.rs"]
 mod communication_tests;
 
-#[path = "software_tests/raspberry_tests.rs"]
-mod raspberry_tests;
+#[path = "software_tests/store_archive.rs"]
+mod store_archive;
+
+#[path = "software_tests/execute_program.rs"]
+mod execute_program;
+
+#[path = "software_tests/stop_program.rs"]
+mod stop_program;
+
+#[path = "software_tests/get_status.rs"]
+mod get_status;
+
+#[path = "software_tests/return_result.rs"]
+mod return_result;
